requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
finnhub-python==2.4.18
fonttools==4.42.1
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
importlib-resources==6.0.1
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.7.2
multidict==6.0.4
numpy==1.25.2
openai==0.27.9
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
psycopg2-binary==2.9.7
PyMySQL==1.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
SQLAlchemy==2.0.20
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
yarl==1.9.2
zipp==3.16.2
finnhub-python